# enhancement: dynamodb performance optimization