<!-- Please refer to the contribution guidelines before raising a PR: https://github.com/localstack/localstack/blob/master/CONTRIBUTING.md -->

<!-- Why am I raising this PR? Add context such as related issues, PRs, or documentation. -->
## Motivation
Ongoing work to optimize DynamoDB performance.

<!-- What notable changes does this PR make? -->
## Changes
- remove the check for Lambda ESM:
  - this is actually not needed, because lambda can only subscribe to a dynamodb stream, [which is what is passed to `CreateEventSourceMapping`](https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Streams.Lambda.Tutorial.html#Streams.Lambda.Tutorial.CreateTrigger). Our check was passing the table ARN, when it should have been the stream ARN, so it never actually returned a result. We have a better in-memory internal way to check if the table has a stream attached, so we should use this (there's actually no public API for DynamoDBStreams).
- optimize the `BatchWriteItem` operation. We would do operations that would have needed to be done only if a stream was attached all the time. Changed those operations to be done only if needed. Refactored the code. (We should still optimize it later on with some batch read, but the priority is performance when no streams are attached).

## Testing

Benchmarks were done and show that an empty provider implementation with only forwarding is as fast as this PR, which provides:
- 1.85x increase in throughput for `PutItem`
- 8.25x increase for `BatchWrite` with batch size = `10`
- 12x increase for `BatchWrite` with batch size = `25`

Small results from the benchmark with LocalStack running in Docker on macOS:
- put 500 items with PutItem = 10.55s -> ~5.5 +-0.1
- put 5000 items with BatchWriteItem = 62s -> 7.6s (to put it in perspective, reducing the time to seed 600k records from 2hrs to 15min)
- put 12500 items with BatchWriteItem = 120s -> 10.2s (which further reduces it from 1h36 to 8min)

## TODO

What's left to do:

- [x] write AWS validated tests to validate the DDB Streams forwarding changes
- [x] write more test around BatchItemWrite to validate behaviour changes


_related issue #1205_